refactor provider unit tests
