+++ plugins/auth-backend/src/providers/pinniped/index.ts
@@ -13,69 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-// import { OidcAuthResult } from '../oidc';
-// import { OidcAuthProvider } from '../oidc/provider';
-// import { OAuthAdapter, OAuthEnvironmentHandler } from '../../lib/oauth';
-// import { createAuthProviderIntegration } from '../createAuthProviderIntegration';
-// import { AuthHandler, SignInResolver } from '../types';
-// import { PinnipedAuthProvider } from './provider';
-
-/**
- * Auth provider integration for Pinniped
- *
- * @public
- */
-// export const pinniped = createAuthProviderIntegration({
-//   create(options?: {
-//     authHandler?: AuthHandler<OidcAuthResult>;
-//     signIn?: {
-//       resolver: SignInResolver<OidcAuthResult>;
-//     };
-//   }) {
-//     return ({ providerId, globalConfig, config, resolverContext }) =>
-//       OAuthEnvironmentHandler.mapConfig(config, envConfig => {
-//         const clientId = envConfig.getString('clientId');
-//         const clientSecret = envConfig.getString('clientSecret');
-//         const customCallbackUrl = envConfig.getOptionalString('callbackUrl');
-//         const callbackUrl =
-//           customCallbackUrl ||
-//           `${globalConfig.baseUrl}/${providerId}/handler/frame`;
-//         const metadataUrl = `${envConfig.getString(
-//           'federationDomain',
-//         )}/.well-known/openid-configuration`;
-//         const federationDomain = envConfig.getString('federationDomain');
-//         const tokenSignedResponseAlg = 'ES256';
-//         const prompt = 'auto';
-//         const authHandler: AuthHandler<OidcAuthResult> = async ({
-//           userinfo,
-//         }) => ({
-//           profile: {},
-//         });
-
-//         // const provider = new OidcAuthProvider({
-//         //   clientId,
-//         //   clientSecret,
-//         //   callbackUrl,
-//         //   tokenSignedResponseAlg,
-//         //   metadataUrl,
-//         //   prompt,
-//         //   signInResolver: options?.signIn?.resolver,
-//         //   authHandler,
-//         //   resolverContext,
-//         // });
-
-//         const provider = new PinnipedAuthProvider({
-//           federationDomain,
-//           clientId,
-//           clientSecret,
-//         });
-
-//         return OAuthAdapter.fromConfig(globalConfig, provider, {
-//           providerId,
-//           callbackUrl,
-//         });
-//       });
-//   },
-// });
 
 export { pinniped } from './provider';
