+++ plugins/auth-backend/src/providers/pinniped/provider.ts
@@ -33,7 +33,6 @@ import { OAuthStartResponse } from '../types';
 import express from 'express';
 import { OAuthAdapter, OAuthEnvironmentHandler } from '../../lib/oauth';
 import { createAuthProviderIntegration } from '../createAuthProviderIntegration';
-import fetch from 'node-fetch';
 
 type OidcImpl = {
   strategy: OidcStrategy<undefined, Client>;
@@ -55,23 +54,15 @@ export type PinnipedOptions = OAuthProviderOptions & {
 
 export class PinnipedAuthProvider implements OAuthHandlers {
   private readonly implementation: Promise<OidcImpl>;
-  private readonly clientId: string;
-  private readonly clientSecret: string;
-  private readonly federationDomain: string;
 
   constructor(options: PinnipedOptions) {
     this.implementation = this.setupStrategy(options);
-    this.clientId = options.clientId;
-    this.clientSecret = options.clientSecret;
-    this.federationDomain = options.federationDomain;
   }
 
   async start(req: OAuthStartRequest): Promise<OAuthStartResponse> {
     const { strategy } = await this.implementation;
-
     const stringifiedAudience = req.query?.audience as string;
     const state = { ...req.state, audience: stringifiedAudience };
-
     const options: Record<string, string> = {
       scope:
         req.scope || 'openid pinniped:request-audience username offline_access',
@@ -88,33 +79,10 @@ export class PinnipedAuthProvider implements OAuthHandlers {
     });
   }
 
-  private async rfc8693TokenExchange({ accessToken, audience }) {
-    const tokenEndpoint = `${this.federationDomain}/oauth2/token`;
-    const authString: string = `${this.clientId}:${this.clientSecret}`;
-    const encodedAuthString = Buffer.from(authString, 'base64');
-
-    const requestOptions = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'x-www-form-urlencoded',
-        Authorization: `Basic ${encodedAuthString}`,
-      },
-      body: `grant_type=urn:ietf:params:oauth:grant-type:token-exchange
-      &subject_token=${accessToken}
-      &subject_token_type=urn:ietf:params:oauth:token-type:access_token
-      &requested_token_type=urn:ietf:params:oauth:token-type:jwt
-      &audience=${audience}`,
-    };
-    const response = await fetch(tokenEndpoint, requestOptions);
-    return response.idToken;
-  }
-
   async handler(
     req: express.Request,
   ): Promise<{ response: OAuthResponse; refreshToken?: string }> {
     const { strategy, client } = await this.implementation;
-
-    // if we dont add a base url our integration fails with invalid_url error in integration test
     const { searchParams } = new URL(req.url, 'https://pinniped.com');
     const stateParam = searchParams.get('state');
     const audience = stateParam ? readState(stateParam).audience : undefined;
@@ -179,7 +147,8 @@ export class PinnipedAuthProvider implements OAuthHandlers {
         response: {
           providerInfo: {
             accessToken: tokenset.access_token!,
-            scope: 'none',
+            scope: tokenset.scope!,
+            idToken: tokenset.id_token,
           },
           profile: {},
         },
